Add files via upload
This is the code I had for the Flask server, I had troubles with formatting data correctly for Foxglove to accept it directly via the websocket so I moved to using something based on the Foxglove websocket example server instead. I have more old Flask stuff that I tried if needed.